Restructure CAFTOP api into a folder with types.ts
