+++ src/Steps/LRDP/LRDP.tsx
@@ -5,7 +5,7 @@ import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import "Steps/Steps.css";
 import { zodResolver } from "@hookform/resolvers/zod";
 
-import { CAFTOPLRDP } from "api/CAFTOP";
+import { CAFTOPLRDP } from "api/CAFTOP/types";
 import { useLRDPPageValidation } from "utilities/Validations";
 import * as Fields from "./Fields";
 import { ICAFTOPWizardStep } from "Steps/Steps";
