+++ src/Steps/Labor/Fields/MILSTD3048.tsx
@@ -1,4 +1,4 @@
-import { CAFTOPLabor } from "api/CAFTOP";
+import { CAFTOPLabor } from "api/CAFTOP/types";
 import BACRadioGroup from "components/BaseFormFields/BACRadioGroup";
 import { Radio, Text } from "@fluentui/react-components";
 import { useWatch } from "react-hook-form";
