+++ src/Steps/Labor/Labor.tsx
@@ -4,7 +4,7 @@ import { globalContext } from "stateManagement/GlobalStore";
 import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import "Steps/Steps.css";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { CAFTOPLabor } from "api/CAFTOP";
+import { CAFTOPLabor } from "api/CAFTOP/types";
 import * as Fields from "./Fields";
 import { useLaborPageValidation } from "utilities/Validations";
 import { ICAFTOPWizardStep } from "Steps/Steps";
