+++ src/Steps/TechnicalOrders/Fields/TOAPMigration.No.tsx
@@ -1,4 +1,4 @@
-import { CAFTOPTechnicalOrders } from "api/CAFTOP";
+import { CAFTOPTechnicalOrders } from "api/CAFTOP/types";
 import BACTextarea from "components/BaseFormFields/BACTextarea";
 import { TOAPMigrationWaiver } from "./TOAPMigration.Waiver";
