+++ src/Steps/TechnicalOrders/TechnicalOrders.tsx
@@ -5,7 +5,7 @@ import { FormProvider, SubmitHandler, useForm } from "react-hook-form";
 import "Steps/Steps.css";
 import { zodResolver } from "@hookform/resolvers/zod";
 
-import { CAFTOPTechnicalOrders } from "api/CAFTOP";
+import { CAFTOPTechnicalOrders } from "api/CAFTOP/types";
 import { useTechnicalOrdersPageValidation } from "utilities/Validations";
 import * as Fields from "./Fields";
 import { ICAFTOPWizardStep } from "Steps/Steps";
